Add files via upload
This file contains the code to generate the China exports matrices to NZL for the year of 1995, 2000 and 2007. Other years can be generated using the same code with otehr datasets in the file. Notice that for each year the code generates two matrices, CN2 and CN3, each indicating to a specific type of China exports. A simple procedure is to add up those two types.